refactor, Extract common test case type for executing tests
